requirements.txt
PyYAML
openai
google-generativeai
python-dotenv
fastapi
uvicorn
python-multipart
Pillow




